[lab6] remove warnings

+++ lab6/Mandelbrot/Mandelbrot/Mandelbrot.cpp
@@ -43,7 +43,7 @@ void Mandelbrot::Draw() const
 	glViewport(0, 0, static_cast<GLsizei>(m_windowSize.width), static_cast<GLsizei>(m_windowSize.height));
 	
 	m_shaderProgram.SetUniform2f(m_mandelbrotUniformMap.systemResolution.c_str(), glm::vec2(m_windowSize.width, m_windowSize.height));
-	m_shaderProgram.SetUniform1f(m_mandelbrotUniformMap.systemTime.c_str(), glfwGetTime());
+	m_shaderProgram.SetUniform1f(m_mandelbrotUniformMap.systemTime.c_str(), static_cast<float>(glfwGetTime()));
 	m_shaderProgram.SetUniform3f(m_mandelbrotUniformMap.mouse.c_str(), glm::vec3(m_mousePosition.x, m_mousePosition.y, m_mousePosition.z));
 
 	m_rectangle.Draw();
